refactor: Processor에서 비즈니스 로직 기능 분리
