+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -5,11 +5,7 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.Socket;
 import java.nio.charset.StandardCharsets;
-import java.util.Map;
-import java.util.Optional;
-import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.UncheckedServletException;
-import nextstep.jwp.model.User;
 import org.apache.coyote.Processor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -36,8 +32,9 @@ public class Http11Processor implements Runnable, Processor {
              final BufferedReader headerReader = new BufferedReader(new InputStreamReader(inputStream,
                      StandardCharsets.UTF_8))) {
             final HttpRequestStartLine startLine = HttpRequestStartLine.from(headerReader);
-
-            checkUser(startLine);
+            
+            final Handler handler = new Handler(startLine);
+            handler.handle();
 
             final HttpResponse httpResponse = HttpResponse.from(startLine);
 
@@ -45,26 +42,6 @@ public class Http11Processor implements Runnable, Processor {
             outputStream.flush();
         } catch (IOException | UncheckedServletException e) {
             log.error(e.getMessage(), e);
-        }
-    }
-
-    private void checkUser(HttpRequestStartLine startLineContents) {
-        if (!startLineContents.getUri().contains("login")) {
-            return;
-        }
-
-        final Map<String, String> queryParams = startLineContents.getQueryParams();
-        final Optional<User> user = InMemoryUserRepository.findByAccount(queryParams.get("account"));
-
-        if (user.isPresent()) {
-            final User loginUser = user.get();
-            logLoginUser(queryParams, loginUser);
-        }
-    }
-
-    private void logLoginUser(Map<String, String> queryParams, User loginUser) {
-        if (loginUser.checkPassword(queryParams.get("password"))) {
-            log.info("user : " + loginUser);
         }
     }
 }
